[flash_ctrl] Make configuration a parameter
To allow for smaller devices the flash size should be
configurable. The PagesPerBank setting is the best candidate to scale
it.

Allow the user to overwrite this value with a define. Also move all
parameter values into the core itself to avoid a dependency on the
toplevel package (makes it usable outside this project too).

Signed-off-by: Stefan Wallentowitz <stefan.wallentowitz@hm.edu>